feat(verify): membuat implementasi untuk verify
