+++ app/api/auth/token/verify/route.ts
@@ -1,15 +1,21 @@
 import { NextResponse } from "next/server";
-import { extractToken } from "../../../utils/authUtils";
+import { extractToken } from "@/app/utils/authUtils";
 import authService from "@/app/services/authService";
 import { ErrorResponse } from "@/app/utils/exceptions";
 
-export async function POST(req: Request) {
+export async function GET(req: Request) {
   const authorization = req.headers.get("Authorization") ?? "";
   try {
     const token = extractToken(authorization);
-    await authService.logout(token);
-    const response = NextResponse.json({ message: "Logged out" });
-    return authService.putRefreshTokenInCookie(response, "");
+    const user = await authService.verify(token);
+
+    return NextResponse.json(
+      {
+        data: { user },
+        message: "Token verified",
+      },
+      { status: 200 }
+    );
   } catch (error) {
     return (error as ErrorResponse).generate();
   }
